Conversion of Binary Event Sequence from double to int.

+++ src/CPP Static Library/SegmentSignal.h
@@ -147,7 +147,7 @@ namespace Algorithms
 			/// <param name="signalLength">Input signal to be segmented.</param>
 			/// <param name="threshold">Length a zone must be to be consider "significant."  Zones shorter than this are ignored.</param>
 			/// <returns>A vector of arrays of length 2 indicating the zones.  Each entry in the List is a pair indicating the starting and ending index of one significant zone.</returns>
-			static vector<array<int, 2>>* FindSignificantZones(double binaryEvents[], double xData[], int signalLength, double threshold);
+			static vector<array<int, 2>>* FindSignificantZones(int binaryEvents[], double xData[], int signalLength, double threshold);
 
 			/// <summary>
 			/// Creates a list of indexes which indicate where the "significant zones" are in the results of the segmentation.  A significant
@@ -166,7 +166,7 @@ namespace Algorithms
 			/// <param name="threshold">Length a zone must be to be consider "significant."  Zones shorter than this are ignored.</param>
 			/// <param name="includeBoundries">If true, the end data points are added as part of the zone.  Does not change how zones are found.</param>
 			/// <returns>A vector of arrays of length 2 indicating the zones.  Each entry in the List is a pair indicating the starting and ending index of one significant zone.</returns>
-			static vector<array<int, 2>>* FindSignificantZones(double binaryEvents[], double xData[], int signalLength, double threshold, bool includeBoundries);
+			static vector<array<int, 2>>* FindSignificantZones(int binaryEvents[], double xData[], int signalLength, double threshold, bool includeBoundries);
 
 		private:
 			/// <summary>
@@ -176,7 +176,7 @@ namespace Algorithms
 			/// <param name="signalLength">Input signal to be segmented.</param>
 			/// <param name="currentIndex">Index to start scanning from.</param>
 			/// <returns>Index of next zero.</returns>
-			static int FindNextZero(double binaryEvents[], int signalLength, int currentIndex);
+			static int FindNextZero(int binaryEvents[], int signalLength, int currentIndex);
 
 			/// <summary>
 			/// Checks to see if the threshold for being a significant zone is met.  If it is, it is added to the list of significant zones.
